dts: msm8996: Add DRD mode for 2nd USB on MSM8996 Auto
Did not bother manually changing references to android_usb1 because we
don't seem to be using it.

Cherry-picked for arm64/[...]/msm8996.dtsi from:

https://github.com/sonyxperiadev/kernel/commit/7b30fb617e47ee6b7efc2af659fffcde7f2789cf
ARM: dts: msm: Add DRD mode for 2nd USB on MSM8996 Auto

Enable DRD mode for 2nd USB. Added one more android_usb
node with usb-core-id=1 so that on the basis of this
usb-core-id we can differentiate between primary and
secondary controller.

Change-Id: I51adc255ce2c3bf4476a490cf3a81fd0512754a8
Signed-off-by: Pratham Pratap <prathampratap@codeaurora.org>