arm64: dts: msm8956-v1.1: Remove 1958MHz node
The msm8956 chip runs at 1.8 GHz in the BIG cluster; running
with a higher clock rate requires overvolting.

This clock has been observed to cause "random" reboots when
set to 1958MHz.

In case verclocking is desired, add 50 ~ 75 mV in the
regulator node of the BIG CPU corresponding to this clock.

Signed-off-by: electronicslover <luizpaulodamaceno@live.com>